feat(contact-service)update the contact-service file and deleted the dupliicate contact file
