--- annuaire/src/app/contact.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { ContactService } from './contact'; 
-
-describe('Contact', () => {
-  let service: ContactService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ContactService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
